requirements.txt
Flask
Flask-Login
Flask-Mail
mysql-connector-python
pandas
openpyxl
Werkzeug
Flask-Bootstrap 